Two conflicts, both where upstream #773 (fixed-width integer types) and
#771 (difficulty_bench) touched what this branch had reshaped:

- lib/Util/RunTimes.cpp: keep upstream's widened free stp::getCurrentTime
  (int64_t, i386-safe) in place of the old member this branch had left
  alone, and keep this branch's processCpuTime/peakMemoryMB helpers --
  outside namespace stp, since RunTimes itself is not in it.
- tools/CMakeLists.txt: keep this branch's per-backend gating for
  rewrite_rule_gen and propagator_bench, and add upstream's new
  difficulty_bench unconditionally (it only scores formulas and works
  with any backend).

Both configurations (USE_MINISAT=ON with CaDiCaL, and CaDiCaL-only)
rebuild and pass the full 104-test suite after the merge; with
BUILD_EXTRA_TOOLS=ON the new difficulty_bench builds and the backend
gating reports as intended.

+++ include/stp/Sat/MinisatCore.h
@@ -71,7 +71,7 @@ public:
 
   void setVerbosity(int v) override;
 
-  unsigned long nVars() const override;
+  uint32_t nVars() const override;
 
   void printStats() const override;
